+++ python/ooi_data_explorations/qartod/qc_processing.py
@@ -560,7 +560,7 @@ def parse_qc(ds):
                 flags[~m, index] = flag  # False == suspect/fail
 
         # add the qc_flags to the dataset, rolling up the results into a single value
-        ds[qc_summary] = ('time', flags.max(axis=1, initial=1))
+        ds[qc_summary] = ('time', flags.max(axis=1, initial=1).astype(np.int32))
 
         # set up the attributes for the new variable
         ds[qc_summary].attrs = dict({
@@ -570,7 +570,7 @@ def parse_qc(ds):
                         'the values are 1 == pass, 2 == not evaluated, 3 == suspect or of high interest, '
                         '4 == fail, and 9 == missing. The QC tests, as applied by OOI, only yield pass or '
                         'fail values.'),
-            'flag_values': np.array([1, 2, 3, 4, 9]),
+            'flag_values': np.array([1, 2, 3, 4, 9]).astype(np.int32),
             'flag_meanings': 'pass not_evaluated suspect_or_of_high_interest fail missing'
         })
